requirements.txt
absl-py==0.9.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cupy-cuda91==7.8.0
Cython==0.29.21
fastrlock==0.5
google-auth==1.21.3
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
importlib-metadata==2.0.0
Markdown==3.2.2
ninja==1.10.0.post2
numpy==1.19.2
oauthlib==3.1.0
pkg-resources==0.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynvrtc==8.0
PyTorch-QRNN @ git+https://github.com/theblackcat102/pytorch-qrnn@10158a41b205d49c9ee9676a0e81004dc4003752
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sru==2.4.3
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
torch==1.4.0
tqdm==4.49.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.2.0
